unify docker and registry models
